# Bugfix and update API handling