--- a/pkg/api/rest/server.go
+++ b/pkg/api/rest/server.go
@@ -202,11 +202,11 @@ func RunServer(port string) {
 	gBeetle.GET("/api/*", echoSwagger.WrapHandler)
 
 	// System management APIs
-	gBeetle.GET("/readyz", rest_common.RestGetReadyz)
-	gBeetle.GET("/httpVersion", rest_common.RestCheckHTTPVersion)
+	gBeetle.GET("/readyz", rest_common.GetReadyz)
+	gBeetle.GET("/httpVersion", rest_common.CheckHTTPVersion)
 
 	// Test utility APIs
-	gBeetle.GET("/test/tracing", rest_common.RestGetTestTracing)
+	gBeetle.GET("/test/tracing", rest_common.TestTracing)
 
 	// Namespace API group
 	// gNamespace := gBeetle.Group("/ns")